requirements.txt
# Core dependencies
opencv-python-headless==4.9.0.80
ultralytics==8.1.0
numpy>=1.24.0

# YouTube stream capture
yt-dlp>=2024.1.0

# Telegram bot
python-telegram-bot==20.7

# Configuration
python-dotenv==1.0.0

# Image processing
Pillow>=10.0.0
